+++ modules/aurora_core/include/events/EventType.h
@@ -5,6 +5,9 @@ enum class EventType {
 	EngineStarted,
 	EngineStopped,
 
+	ModuleLoaded,
+	ModuleUnloaded,
+
 	WindowResize,
 	WindowClose,