+++ modules/aurora_core/include/events/engine/EngineStoppedEvent.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "../EventMacros.h"
+#include "../IEvent.h"
+
+class EngineStoppedEvent : public IEvent {
+public:
+	EVENT_CLASS_TYPE(EngineStopped)
+	EVENT_CLASS_CATEGORY(Engine)
+};